+++ firmware/hdl/payload/frame.vhd
@@ -1,7 +1,7 @@
 -- Description:
 -- Contains the "framework" of GT-logic (all parts, except GTL and FDL).
 
--- HB 2022-03-22: v1.3.0 - output ports bcres_d and bcres_d_FDL not used anymore (not used in mp7_payload.vhd). Signals bcres and bcres_outputmux not used anymore. Updated tcm.vhd (input port bcres_d_FDL not used anymore).
+-- HB 2022-03-22: v1.3.0 - output ports bcres_d and bcres_d_FDL not used anymore (not used in mp7_payload.vhd). Signals bcres, bcres_outputmux, bcres_d_FDL_int and bx_nr_d_FDL_int not used anymore. Updated tcm.vhd (input port bcres_d_FDL not used anymore) and output_mux.vhd (input port bx_nr_fdl not used anymore). Removed signals for spy3.
 -- HB 2021-06-16: v1.2.4 - implemented selectors (set in gtl_pkg.vhd) for "scouting" (in output_mux.vhd) and use of input data spymem.
 -- HB 2017-10-10: v1.2.3 - bug fix "simmem_in_use_i" input of spytrig.
 -- HB 2017-10-10: v1.2.2 - removed mux control register ("mux_ctrl_regs_1"), used fixed values for output mux inputs ("mux_ctrl").
@@ -63,8 +63,6 @@ entity frame is
         oc0 : in std_logic;
         start : in std_logic;
         l1a : in std_logic;
---         bcres_d : out std_logic;
---         bcres_d_FDL : out std_logic;
         start_lumisection : out std_logic;
         lane_data_in : in ldata(NR_LANES-1 downto 0);
         lane_data_out : out ldata(NR_LANES-1 downto 0);
@@ -106,21 +104,15 @@ architecture rtl of frame is
 
     signal lmp_lhc_data_o   : lhc_data_t; -- lhc_data output of lane mapping process
 
---     signal bcres : std_logic; -- NOT USED, "bc0" of mp7_ttc is used instead of "bcres"
     signal bc0_d_int : std_logic; -- delayed version of bcres
---     signal bcres_d_FDL_int : std_logic; -- delayed version of bcres for FDL
---     signal bcres_outputmux : std_logic; -- non-delayed version of bcres for output mux
 
     --TCM signals
     signal bx_nr : bx_nr_t;
-    signal bx_nr_d_FDL_int : bx_nr_t;
     signal orbit_nr : orbit_nr_t;
 
     -- sim/spy mem
     signal spy1 : std_logic;
     signal spy2 : std_logic;
-    signal spy3 : std_logic;
-    signal spy3_ack : std_logic;
 
     signal lhc_data_slv_o : std_logic_vector(LHC_DATA_WIDTH-1 downto 0);
     signal lhc_data_slv_i : std_logic_vector(LHC_DATA_WIDTH-1 downto 0);
@@ -223,11 +215,9 @@ architecture rtl of frame is
             start             => start_d_int,
             l1a_sync          => l1a,
             bcres_d           => bc0_d_int,
---             bcres_d_FDL       => bcres_d_FDL_int,
             sw_reg_in         => rb2tcm,
             sw_reg_out        => tcm2rb,
             bx_nr             => bx_nr,
-            bx_nr_d_FDL       => bx_nr_d_FDL_int,
             event_nr          => open,
             trigger_nr        => open,
             orbit_nr          => orbit_nr,
@@ -264,25 +254,16 @@ architecture rtl of frame is
     begin
         if lhc_rst = RST_ACT then
             bc0_d_int <= '0';
---             bcres_d_int <= '0';
             ec0_d_int <= '0';
             oc0_d_int <= '0';
             start_d_int <= '0';
---             bcres_d_FDL_int <= '0';
---             bcres_outputmux <= '0';
         elsif rising_edge(lhc_clk) then
             bc0_d_int  <= bc0;
---             bcres_d_int  <= bc0;
             ec0_d_int <= ec0;
             oc0_d_int <= oc0;
             start_d_int <= start;
---             bcres_d_FDL_int <= bc0;
---             bcres_outputmux <= bc0;
         end if;
     end process;
-
---     bcres_d_FDL <= bcres_d_FDL_int;
---     bcres_d <= bcres_d_int;
 
 --===============================================================================================--
 --                           SIM/SPY MEMORY
@@ -296,13 +277,8 @@ architecture rtl of frame is
             bx_nr      => bx_nr,
             sw_reg_i   => rb2spytrig,
             sw_reg_o   => spytrig2rb,
-
             spy1_o     => spy1,
-            spy2_o     => spy2,
-            spy3_o     => spy3,
-            spy3_ack_i => spy3_ack,
-
-            simmem_in_use_i => '0'
+            spy2_o     => spy2
         );
 
 -- use of spymem depends on selector SPYMEM (set in gtl_pkg.vhd)
@@ -397,7 +373,7 @@ architecture rtl of frame is
             lhc_rst     => lhc_rst,
             ctrs        => ctrs,
             bx_nr       => bx_nr,
-            bx_nr_fdl   => bx_nr_d_FDL_int,
+--             bx_nr_fdl   => bx_nr_d_FDL_int,
             orbit_nr    => orbit_nr,
             algo_after_gtLogic   => algo_after_gtLogic_rop,
             algo_after_bxomask   => algo_after_bxomask_rop,